pyproject.toml
[project]
name = "aoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "typer>=0.13.1",
]
